Refactor BadgeInstance update tasks and command
